+++ src/ielts/tools/push.py
@@ -1,1 +1,0 @@
-"""Push notification wrapper. To be implemented in P0-07."""